Autoload custom stuff + slight cleanup
Here "cleanup" mostly means "fix indentation/layout", but in the case of
`omnisharp-server-executable-path' I actually had to go so far as to add
a :group keyword to get it into the custom group with the rest.

The main goal here is to help users find the settings they know they'll
need to fix *before* they can expect `omnisharp-mode' to work,
principally `omnisharp-server-executable-path'.